OCRAdhar/requirements.txt
pillow
pytesseract
Aadhaar-extractor
requests